TestingPage_Backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
# File: backend/requirements-complete.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
matplotlib==3.8.2
networkx==3.2
google-generativeai
python-dotenv